// frontend/src/generated/model/messages.ts
/**
 * api-gateway
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Messages { 
    id?: any;
    conversations_id?: any;
    sender?: any;
    message?: any;
    date?: any;
}