// frontend/src/generated/model/conversations.ts
/**
 * api-gateway
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Conversations { 
    id?: any;
    user_id?: any;
    model_id?: any;
    title?: any;
    created_at?: any;
    updated_at?: any;
}